pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "buddy-mk67"
version = "0.2.0"
description = "Windows-first autonomous desktop assistant with supervised runtime execution."
readme = "readme.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [
  { name = "Sirius" }
]
dependencies = [
  "PyQt6",
  "beautifulsoup4",
  "chromadb",
  "comtypes; platform_system == 'Windows'",
  "docx2txt",
  "duckduckgo-search",
  "google-genai",
  "google-generativeai",
  "mss",
  "numpy",
  "opencv-python",
  "pillow",
  "playwright",
  "psutil",
  "pyautogui",
  "pyaudio",
  "pycaw; platform_system == 'Windows'",
  "pygetwindow",
  "pyperclip",
  "pywinauto; platform_system == 'Windows'",
  "requests>=2.33.1,<3",
  "send2trash",
  "sentence-transformers",
  "sounddevice",
  "youtube-transcript-api",
  "python-telegram-bot>=21.0",
  "keyboard"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23"
]

[project.scripts]
buddy = "main:main"
buddy-ship-check = "scripts.ship_ready_check:main"

[tool.setuptools]
py-modules = ["main", "ui", "app_bootstrap", "buddy_logging"]

[tool.setuptools.packages.find]
include = ["actions*", "agent*", "config*", "memory*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
